+++ core/src/main/java/io/github/bekoenig/getdown/data/SysProps.java
@@ -135,6 +135,14 @@ public final class SysProps {
     }
 
     /**
+     * If false, Getdown will not detect proxy or request proxy info. Default is true.
+     * Usage: {@code -Duse_proxy=false}.
+     */
+    public static boolean useProxy() {
+        return !"false".equals(System.getProperty("use_proxy"));
+    }
+
+    /**
      * If true, Getdown will always try to connect without proxy settings even it a proxy is set
      * in {@code proxy.txt}. If direct access is possible it will not clear {@code proxy.txt}, it
      * will preserve the settings. This is to support cases where a user uses a workstation in two